Add intermediate model classes
