+++ pybop/models/empirical/ecm_base.py
@@ -1,0 +1,30 @@
+import pybamm
+from ..base_model import BaseModel
+
+
+class ECircuitModel(BaseModel):
+    """
+    An intermediate class to extend the functionality of the electrochemical PyBaMM models.
+    """
+
+    def __init__(self):
+        super().__init__()
+
+    def _check_params(self, inputs=None, allow_infeasible_solutions=True):
+        """
+        Check the compatibility of the model parameters.
+
+        Parameters
+        ----------
+        inputs : dict
+            The input parameters for the simulation.
+        allow_infeasible_solutions : bool, optional
+            If True, infeasible parameter values will be allowed in the optimisation (default: True).
+
+        Returns
+        -------
+        bool
+            A boolean which signifies whether the parameters are compatible.
+
+        """
+        return True
